refactor(client): /resultのコンポーネントを/result/[parent_id]へ移動
