+++ client/src/app/result/[parent_id]/_components/copy-button/copy-button.tsx
