+++ client/src/app/result/[parent_id]/_components/copy-button/index.ts
