+++ client/src/app/result/[parent_id]/_components/create-link/create-link.tsx
