+++ client/src/app/result/[parent_id]/_components/index.ts
