+++ client/src/app/result/[parent_id]/_components/misskey-share-button/index.ts
