+++ client/src/app/result/[parent_id]/_components/modal/modal.css.ts
