+++ client/src/app/result/[parent_id]/_components/parents/index.ts
