+++ client/src/app/result/[parent_id]/_components/parents/parents.tsx
