+++ client/src/app/result/[parent_id]/_components/result-image/index.ts
