+++ client/src/app/result/[parent_id]/_components/save-link/index.ts
