+++ client/src/app/result/[parent_id]/_components/title/title.tsx
