+++ client/src/app/result/[parent_id]/page.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { Button } from "@/ui";
 import {
   CopyButton,
   MisskeyShareButton,
@@ -10,13 +9,13 @@ import {
   SaveLink,
   TwitterShareButton,
   Title,
-} from "../_components";
+} from "./_components";
 import { useEffect, useRef, useState } from "react";
 import { TMorphoto } from "@/types/Morphoto";
 import { createInference, readGcs, readMorphoto, readStatus } from "@/api";
 import { env } from "@/constants";
 
-import * as styles from "../result.css";
+import * as styles from "./result.css";
 
 export default function ResultPage({
   params: { parent_id },
@@ -25,7 +24,6 @@ export default function ResultPage({
   params: { parent_id: string };
   searchParams: { prompt?: string; strength?: string };
 }) {
-  const [modalOpen, setModalOpen] = useState<boolean>(false);
   const [morphoto, setMorphoto] = useState<TMorphoto | null>(null);
   const [done, setDone] = useState<boolean>(false);
   const [gcsRegistered, setGcsRegistered] = useState<boolean>(true);
@@ -119,7 +117,6 @@ export default function ResultPage({
 
   useEffect(() => {
     checkGcs();
-    setModalOpen(true);
     checkStatus();
     morphing();
     getMorphoto();
