+++ client/src/app/result/[parent_id]/result.css.ts
